Determine with 'ftp:/' instead of QUrl::scheme()
QUrl::scheme() thinks that a scheme of 'ftp:' is 'ftp'. And 'ftp:/'
should be used not 'ftp://'. This is due to a bug of QFileSystemModel
not understanding URL scheme correctly. It always tries to treat all the
things as local paths. So it converts ftp://xxx to ftp:/xxx.

As a side effect, there is no need to check that a directory is 'ftp:'.