Improve cache hits
If cdUp is performed after entering into an empty directory, cache fails
even if it has entires of the parent.

This is because an empty directory is not maintained by cache. Whenever
accessing the empty directory, cache always fails.

To avoid this problem, insert an invalid empty entry at first.