KFileWizard: execute creating QDir() instance in a separate thread
Creating QDir() instance itself blocks a thread. Because of this, a
message for checking accessibility when connecting to SFTP servers, does
not appear. However, when connecting FTP servers, FtpFileEngines uses
internal event loops for QFtp. Due to this, a message may be shown.

To fix this, execute creating QDir() instance in a separate thread, as
well as calling QDir::isReadable() method.