FtpFilEngine: fix regression of commit 37af8f
Commit 37af8fb1d31059e1a0fff2b03c6698520cdc5284 introduces a regression
that FTP listing does not work.
This is because QtConcurrent::run() executes a function in a thread
pool. And QFtp is based on events.

When QFtp instance is created at first, its thread context is a thread
pool created by QtConcurrent::run(). However, that thread is terminated
after executing a function. So events of QFtp are not processed.

To fix this, create/delete QFtp instances in a separate thread living
during a lifetime of a program.